Wicket Metrics - Moved to experimental / doc improvements / aspects
