+++ pom.xml
@@ -114,7 +114,6 @@
 		<module>wicket-bean-validation</module>
 		<module>wicket-cdi-1.1</module>
 		<module>wicket-user-guide</module>
-		<module>wicket-metrics</module>
 	</modules>
 	<properties>
 		<!-- Encoding -->
@@ -406,13 +405,13 @@
 				<type>jar</type>
 			</dependency>
 			<dependency>
-				<groupId>org.apache.wicket</groupId>
+				<groupId>org.apache.wicket.experimental.wicket8</groupId>
 				<artifactId>wicket-metrics</artifactId>
 				<version>8.0.0-SNAPSHOT</version>
 				<type>jar</type>
 			</dependency>
 			<dependency>
-				<groupId>org.apache.wicket.experimental.wicket7</groupId>
+				<groupId>org.apache.wicket.experimental.wicket8</groupId>
 				<artifactId>wicket-atmosphere</artifactId>
 				<version>0.24-SNAPSHOT</version>
 				<type>jar</type>
