+++ wicket-experimental/pom.xml
@@ -23,7 +23,7 @@
 		<version>8.0.0-SNAPSHOT</version>
 		<relativePath>../pom.xml</relativePath>
 	</parent>
-	<groupId>org.apache.wicket.experimental.wicket7</groupId>
+	<groupId>org.apache.wicket.experimental.wicket8</groupId>
 	<artifactId>wicket-experimental</artifactId>
 	<packaging>pom</packaging>
 	<name>Wicket-Experimental</name>
