+++ wicket-user-guide/src/docs/guide/toc.yml
@@ -224,10 +224,11 @@ wicketstuff:
   wicketstuff_6: Module wicketstuff-rest-annotations
   wicketstuff_7: Module stateless
 monitoring:
-  title: Wicket Metrics Monitoring
+  title: Wicket Metrics Monitoring (Experimental)
   monitoring_1: Example setup
   monitoring_2: Visualization with Graphite
   monitoring_3: Measured data
+  monitoring_4: Write own measurements
 redirects:
   title: Lost In Redirection With Apache Wicket (Appendix)
 contributing:
